requirements.txt
fastapi==0.75.0
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.0.1
torch==1.11.0
torchvision==0.12.0
uvicorn==0.17.6
